sql: remove heuristic planner optimization passes
Removing the "optimization" part of the heuristic planner, and any
code that becomes unused.

We still keep the limit propagation pass; it will go away when the
optimizer has intelligence around soft limits.

A small fix that happened naturally is to propagate limits on
`EXPLAIN` which more accurately reflects the actual execution.

Release note: None
