+++ pkg/sql/index_join.go
@@ -125,100 +125,6 @@ func processIndexJoinColumns(
 	return primaryKeyColumns, colIDtoRowIndex
 }
 
-// makeIndexJoin build an index join node.
-// This destroys the original table scan node argument and reuses its
-// storage to construct a new index scan node. A new table scan node
-// is created separately as a member of the resulting index join node.
-// The new index scan node is also returned alongside the new index join
-// node.
-func (p *planner) makeIndexJoin(
-	origScan *scanNode, exactPrefix int,
-) (resultPlan *indexJoinNode, indexScan *scanNode, _ error) {
-	// Reuse the input argument's scanNode and its initialized parameters
-	// at a starting point to build the new indexScan node.
-	indexScan = origScan
-
-	// Create a new scanNode that will be used with the primary index.
-	table := p.Scan()
-	table.desc = origScan.desc
-	// Note: initDescDefaults can only error out if its 3rd argument is not nil.
-	if err := table.initDescDefaults(p.curPlan.deps, origScan.colCfg); err != nil {
-		return nil, nil, err
-	}
-	table.initOrdering(0 /* exactPrefix */, p.EvalContext())
-	table.disableBatchLimit()
-
-	primaryKeyColumns, colIDtoRowIndex := processIndexJoinColumns(table, indexScan)
-
-	// To split the WHERE filter into an index-specific part and a
-	// "rest" part below the splitFilter() code must know which columns
-	// are provided by the index scanNode. Since primaryKeyColumns only
-	// contains the PK columns of the indexed table, we also need to
-	// gather here which additional columns are indexed. This is done in
-	// valProvidedIndex.
-	valProvidedIndex := make([]bool, len(origScan.cols))
-
-	// Then, in case the index-specific part, post-split, actually
-	// refers to any additional column, we also need to prepare the
-	// mapping for these columns in colIDtoRowIndex.
-	if indexScan.index.Type == sqlbase.IndexDescriptor_FORWARD {
-		for _, colID := range indexScan.index.ColumnIDs {
-			idx, ok := indexScan.colIdxMap[colID]
-			if !ok {
-				panic(fmt.Sprintf("Unknown column %d in index!", colID))
-			}
-			valProvidedIndex[idx] = true
-			colIDtoRowIndex[colID] = idx
-		}
-	}
-
-	if origScan.filter != nil {
-		// Now we split the filter by extracting the part that can be
-		// evaluated using just the index columns.
-		splitFunc := func(expr tree.VariableExpr) (ok bool, newExpr tree.Expr) {
-			colIdx := expr.(*tree.IndexedVar).Idx
-			if !(primaryKeyColumns[colIdx] || valProvidedIndex[colIdx]) {
-				return false, nil
-			}
-			return true, indexScan.filterVars.IndexedVar(colIdx)
-		}
-		indexScan.filter, table.filter = splitFilter(origScan.filter, splitFunc)
-	}
-
-	// splitFilter above may have simplified the filter expression and
-	// eliminated all remaining references to some of the
-	// IndexedVars. Rebind the indexed vars here so that these stale
-	// references are eliminated from the filterVars helper and the set
-	// of needed columns is properly determined later by
-	// setNeededColumns().
-	indexScan.filter = indexScan.filterVars.Rebind(indexScan.filter, true, false)
-
-	// Ensure that the remaining indexed vars are transferred to the
-	// table scanNode fully.
-	table.filter = table.filterVars.Rebind(table.filter, true /* alsoReset */, false /* normalizeToNonNil */)
-	indexScan.initOrdering(exactPrefix, p.EvalContext())
-
-	primaryKeyPrefix := roachpb.Key(sqlbase.MakeIndexKeyPrefix(table.desc.TableDesc(), table.index.ID))
-	node := &indexJoinNode{
-		index:             indexScan,
-		table:             table,
-		cols:              table.cols,
-		resultColumns:     table.resultColumns,
-		primaryKeyColumns: primaryKeyColumns,
-		run: indexJoinRun{
-			primaryKeyPrefix: primaryKeyPrefix,
-			colIDtoRowIndex:  colIDtoRowIndex,
-		},
-		// We need to store the physical props in this node because they might be
-		// different from the underlying index scan in plans produced by the
-		// optimizer (though they will be the same in plans produced by the
-		// heuristic planner).
-		props: planPhysicalProps(indexScan),
-	}
-
-	return node, indexScan, nil
-}
-
 // indexJoinRun stores the run-time state of indexJoinNode during local execution.
 type indexJoinRun struct {
 	// primaryKeyPrefix is the KV key prefix of the rows
