+++ pkg/sql/join_predicate.go
@@ -29,9 +29,6 @@ type joinPredicate struct {
 
 	// left/rightEqualityIndices give the position of equality columns
 	// on the left and right input row arrays, respectively.
-	// Left/right columns that have an equality constraint in the ON
-	// condition also have their indices appended when tryAddEqualityFilter
-	// is invoked (see planner.addJoinFilter).
 	// Only columns with the same left and right value types can be equality
 	// columns.
 	leftEqualityIndices  []int
@@ -70,69 +67,6 @@ type joinPredicate struct {
 	// will link to it by reference after checkRenderStar / analyzeExpr.
 	// Enforce this using NoCopy.
 	_ util.NoCopy
-}
-
-// tryAddEqualityFilter attempts to turn the given filter expression into
-// an equality predicate. It returns true iff the transformation succeeds.
-func (p *joinPredicate) tryAddEqualityFilter(
-	filter tree.Expr, left, right *sqlbase.DataSourceInfo,
-) bool {
-	c, ok := filter.(*tree.ComparisonExpr)
-	if !ok || c.Operator != tree.EQ {
-		return false
-	}
-	lhs, ok := c.Left.(*tree.IndexedVar)
-	if !ok {
-		return false
-	}
-	rhs, ok := c.Right.(*tree.IndexedVar)
-	if !ok {
-		return false
-	}
-
-	sourceBoundary := len(left.SourceColumns)
-	if (lhs.Idx >= sourceBoundary && rhs.Idx >= sourceBoundary) ||
-		(lhs.Idx < sourceBoundary && rhs.Idx < sourceBoundary) {
-		// Both variables are on the same side of the join (e.g. `a JOIN b ON a.x = a.y`).
-		return false
-	}
-
-	if lhs.Idx > rhs.Idx {
-		lhs, rhs = rhs, lhs
-	}
-
-	if !lhs.ResolvedType().Equivalent(rhs.ResolvedType()) {
-		// Issue #22519: we can't have two equality columns of mismatched types
-		// because the hash-joiner assumes the encodings are the same.
-		return false
-	}
-
-	// At this point we have an equality, so we can add it to the list
-	// of equality columns.
-
-	// To do this we must be a bit careful: the expression contains
-	// IndexedVars, and the column indices at this point will refer to
-	// the full column set of the joinPredicate, including the
-	// merged columns.
-	p.addEquality(left, lhs.Idx, right, rhs.Idx-len(left.SourceColumns))
-	return true
-}
-
-func (p *joinPredicate) addEquality(
-	left *sqlbase.DataSourceInfo, leftColIdx int, right *sqlbase.DataSourceInfo, rightColIdx int,
-) {
-	// Also, we will want to avoid redundant equality checks.
-	for i := range p.leftEqualityIndices {
-		if p.leftEqualityIndices[i] == leftColIdx && p.rightEqualityIndices[i] == rightColIdx {
-			// The filter is already there.
-			return
-		}
-	}
-
-	p.leftEqualityIndices = append(p.leftEqualityIndices, leftColIdx)
-	p.rightEqualityIndices = append(p.rightEqualityIndices, rightColIdx)
-	p.leftColNames = append(p.leftColNames, tree.Name(left.SourceColumns[leftColIdx].Name))
-	p.rightColNames = append(p.rightColNames, tree.Name(right.SourceColumns[rightColIdx].Name))
 }
 
 // makePredicate constructs a joinPredicate object for joins. The join condition
@@ -280,32 +214,6 @@ func (p *joinPredicate) eval(ctx *tree.EvalContext, leftRow, rightRow tree.Datum
 	return true, nil
 }
 
-// getNeededColumns figures out the columns needed for the two
-// sources.  This takes into account both the equality columns and the
-// predicate expression.
-func (p *joinPredicate) getNeededColumns(neededJoined []bool) ([]bool, []bool) {
-	// Reset the helper and rebind the variable to detect which columns
-	// are effectively needed.
-	p.onCond = p.iVarHelper.Rebind(p.onCond, true, false)
-
-	// The columns that are part of the expression are always needed.
-	neededJoined = append([]bool(nil), neededJoined...)
-	for i := range neededJoined {
-		if p.iVarHelper.IndexedVarUsed(i) {
-			neededJoined[i] = true
-		}
-	}
-	leftNeeded := neededJoined[:p.numLeftCols]
-	rightNeeded := neededJoined[p.numLeftCols:]
-
-	// The equality columns are always needed.
-	for i := range p.leftEqualityIndices {
-		leftNeeded[p.leftEqualityIndices[i]] = true
-		rightNeeded[p.rightEqualityIndices[i]] = true
-	}
-	return leftNeeded, rightNeeded
-}
-
 // prepareRow prepares the output row by combining values from the
 // input data sources.
 func (p *joinPredicate) prepareRow(result, leftRow, rightRow tree.Datums) {
@@ -386,3 +294,17 @@ func pickUsingColumn(
 }
 
 const invalidColIdx = -1
+
+// mergeConj combines two predicates.
+func mergeConj(left, right tree.TypedExpr) tree.TypedExpr {
+	if isFilterTrue(left) {
+		if right == tree.DBoolTrue {
+			return nil
+		}
+		return right
+	}
+	if isFilterTrue(right) {
+		return left
+	}
+	return tree.NewTypedAndExpr(left, right)
+}
