+++ pkg/sql/opt_index_selection.go
@@ -11,528 +11,13 @@
 package sql
 
 import (
-	"context"
-	"fmt"
-	"sort"
-
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/constraint"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/exec"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/exec/execbuilder"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/idxconstraint"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/optbuilder"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/xform"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/errors"
 )
-
-const nonCoveringIndexPenalty = 10
-
-// analyzeOrderingFn is the interface through which the index selection code
-// discovers how useful is the ordering provided by a certain index. The higher
-// layer (select) desires a certain ordering on a number of columns; it calls
-// into the index selection code with an analyzeOrderingFn that computes how
-// many columns of that desired ordering are satisfied by the index ordering.
-// Both the number of matching columns and the total columns in the desired
-// ordering are returned.
-//
-// For example, consider the table t {
-//    a INT,
-//    b INT,
-//    c INT,
-//    INDEX ab (a, b)
-//    INDEX bac (b, a, c)
-// }
-//
-// For `SELECT * FROM t ORDER BY a, c`, the desired ordering is (a, c);
-// totalCols is 2. In this case:
-//  - the primary index has no ordering on a, b, c; matchingCols is 0.
-//  - the ab index matches the first column of the desired ordering;
-//    matchingCols is 1.
-//  - the bac index doesn't match the desired ordering at all; mathcingCols
-//    is 0.
-//
-// For `SELECT * FROM t WHERE b=1 ORDER BY a, c`, the desired ordering is (a, c);
-// totalCols is 2. In this case:
-//  - the primary index has no ordering on a, b, c; matchingCols is 0.
-//  - the ab index matches the first column of the desired ordering;
-//    matchingCols is 1.
-//  - the bac index, along with the fact that b is constrained to a single
-//    value, matches the desired ordering; matchingCols is 2.
-type analyzeOrderingFn func(indexProps physicalProps) (matchingCols, totalCols int)
-
-// selectIndex analyzes the scanNode to determine if there is an index
-// available that can fulfill the query with a more restrictive scan.
-//
-// Analysis currently consists of a simplification of the filter expression,
-// replacing expressions which are not usable by indexes by "true". The
-// simplified expression is then considered for each index and a set of range
-// constraints is created for the index. The candidate indexes are ranked using
-// these constraints and the best index is selected. The constraints are then
-// transformed into a set of spans to scan within the index.
-//
-// The analyzeOrdering function is used to determine how useful the ordering of
-// an index is. If no particular ordering is desired, it can be nil.
-//
-// If preferOrderMatching is true, we prefer an index that matches the desired
-// ordering completely, even if it is not a covering index.
-func (p *planner) selectIndex(
-	ctx context.Context, s *scanNode, analyzeOrdering analyzeOrderingFn, preferOrderMatching bool,
-) (planNode, error) {
-	if s.desc.IsEmpty() {
-		// No table.
-		s.initOrdering(0 /* exactPrefix */, p.EvalContext())
-		return s, nil
-	}
-
-	if s.filter == nil && analyzeOrdering == nil && s.specifiedIndex == nil {
-		// No where-clause, no ordering, and no specified index.
-		s.initOrdering(0 /* exactPrefix */, p.EvalContext())
-		var err error
-		s.spans, err = unconstrainedSpans(s.desc, s.index, s.isDeleteSource)
-		if err != nil {
-			return nil, errors.NewAssertionErrorWithWrappedErrf(err,
-				"table ID = %d, index ID = %d", errors.Safe(s.desc.ID), errors.Safe(s.index.ID))
-		}
-		return s, nil
-	}
-
-	candidates := make([]*indexInfo, 0, len(s.desc.Indexes)+1)
-	if s.specifiedIndex != nil {
-		// An explicit secondary index was requested. Only add it to the candidate
-		// indexes list.
-		candidates = append(candidates, &indexInfo{
-			desc:    s.desc,
-			index:   s.specifiedIndex,
-			reverse: s.specifiedIndexReverse,
-		})
-	} else {
-		candidates = append(candidates, &indexInfo{
-			desc:  s.desc,
-			index: &s.desc.PrimaryIndex,
-		})
-		for i := range s.desc.Indexes {
-			candidates = append(candidates, &indexInfo{
-				desc:  s.desc,
-				index: &s.desc.Indexes[i],
-			})
-		}
-	}
-
-	for _, c := range candidates {
-		c.init(s)
-	}
-
-	var optimizer xform.Optimizer
-
-	if s.filter != nil {
-		optimizer.Init(p.EvalContext())
-		md := optimizer.Memo().Metadata()
-		for i := range s.resultColumns {
-			md.AddColumn(s.resultColumns[i].Name, s.resultColumns[i].Typ)
-		}
-		bld := optbuilder.NewScalar(ctx, &p.semaCtx, p.EvalContext(), optimizer.Factory())
-		bld.AllowUnsupportedExpr = true
-		err := bld.Build(s.filter)
-		if err != nil {
-			return nil, err
-		}
-		filters := memo.FiltersExpr{{Condition: optimizer.Memo().RootExpr().(opt.ScalarExpr)}}
-		filters = optimizer.Factory().CustomFuncs().SimplifyFilters(filters)
-		for _, c := range candidates {
-			if err := c.makeIndexConstraints(
-				&optimizer, filters, p.EvalContext(),
-			); err != nil {
-				return nil, err
-			}
-			if c.ic.Constraint().IsContradiction() {
-				// No spans (i.e. the filter is always false). Note that if a filter
-				// results in no constraints, ok would be false.
-				return newZeroNode(s.resultColumns), nil
-			}
-		}
-	}
-
-	// Remove any inverted indexes that don't generate any spans, a full-scan of
-	// an inverted index is always invalid.
-	for i := 0; i < len(candidates); {
-		c := candidates[i].ic.Constraint()
-		if candidates[i].index.Type == sqlbase.IndexDescriptor_INVERTED && (c == nil || c.IsUnconstrained()) {
-			candidates[i] = candidates[len(candidates)-1]
-			candidates = candidates[:len(candidates)-1]
-		} else {
-			i++
-		}
-	}
-
-	if len(candidates) == 0 {
-		// The primary index is never inverted. So the only way this can happen is
-		// if we had a specified index.
-		if s.specifiedIndex == nil {
-			panic("no non-inverted indexes")
-		}
-		return nil, fmt.Errorf("index \"%s\" is inverted and cannot be used for this query",
-			s.specifiedIndex.Name)
-	}
-
-	if s.noIndexJoin {
-		// Eliminate non-covering indexes. We do this after the check above for
-		// constant false filter.
-		for i := 0; i < len(candidates); {
-			if !candidates[i].covering {
-				candidates[i] = candidates[len(candidates)-1]
-				candidates = candidates[:len(candidates)-1]
-			} else {
-				i++
-			}
-		}
-		if len(candidates) == 0 {
-			// The primary index is always covering. So the only way this can
-			// happen is if we had a specified index.
-			if s.specifiedIndex == nil {
-				panic("no covering indexes")
-			}
-			return nil, fmt.Errorf("index \"%s\" is not covering and NO_INDEX_JOIN was specified",
-				s.specifiedIndex.Name)
-		}
-	}
-
-	for _, c := range candidates {
-		if analyzeOrdering != nil {
-			c.analyzeOrdering(ctx, s, analyzeOrdering, preferOrderMatching, p.EvalContext())
-		}
-	}
-
-	indexInfoByCost(candidates).Sort()
-
-	if log.V(2) {
-		for i, c := range candidates {
-			log.Infof(
-				ctx, "%d: selectIndex(%s): cost=%v constraint=%s reverse=%t",
-				i, c.index.Name, c.cost, c.ic.Constraint(), c.reverse,
-			)
-		}
-	}
-
-	// After sorting, candidates[0] contains the best index. Copy its info into
-	// the scanNode.
-	c := candidates[0]
-	s.index = c.index
-	s.specifiedIndex = nil
-	s.isSecondaryIndex = (c.index != &s.desc.PrimaryIndex)
-
-	constraint := c.ic.Constraint()
-
-	var err error
-	s.spans, err = spansFromConstraint(
-		s.desc, c.index, constraint, s.valNeededForCol, s.isDeleteSource)
-	if err != nil {
-		return nil, errors.NewAssertionErrorWithWrappedErrf(
-			err, "constraint = %s, table ID = %d, index ID = %d",
-			constraint, errors.Safe(s.desc.ID), errors.Safe(s.index.ID),
-		)
-	}
-
-	if len(s.spans) == 0 {
-		// There are no spans to scan.
-		return newZeroNode(s.resultColumns), nil
-	}
-
-	s.origFilter = s.filter
-	if s.filter != nil {
-		rem := c.ic.RemainingFilters()
-		if rem.IsTrue() {
-			s.filter = nil
-		} else {
-			execBld := execbuilder.New(nil /* execFactory */, optimizer.Memo(), nil /* catalog */, &rem, nil /* evalCtx */)
-			s.filter, err = execBld.BuildScalar(&s.filterVars)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	s.filterVars.Rebind(s.filter, true, false)
-
-	s.reverse = c.reverse
-	s.maxResults = calculateMaxResults(constraint, c.desc, c.index, p.EvalContext())
-
-	var plan planNode
-	if c.covering {
-		s.initOrdering(c.exactPrefix, p.EvalContext())
-		plan = s
-	} else {
-		// Note: makeIndexJoin destroys s and returns a new index scan
-		// node. The filter in that node may be different from the
-		// original table filter.
-		plan, s, err = p.makeIndexJoin(s, c.exactPrefix)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if log.V(3) {
-		log.Infof(ctx, "%s: filter=%v", c.index.Name, s.filter)
-		for i, span := range s.spans {
-			log.Infof(ctx, "%s/%d: %s", c.index.Name, i, sqlbase.PrettySpan(sqlbase.IndexKeyValDirs(c.index), span, 2))
-		}
-	}
-
-	return plan, nil
-}
-
-// calculateMaxResults returns the maximum number of results for a scan; the
-// scan is guaranteed never to return more results than this. Iff this hint is
-// invalid, 0 is returned.
-// This is a duplicate of the method Builder.indexConstraintMaxResults in the
-// execbuilder package - this exists so that the heuristic planner, which plans
-// mutations, can still detect this condition.
-func calculateMaxResults(
-	c *constraint.Constraint,
-	t *sqlbase.ImmutableTableDescriptor,
-	i *sqlbase.IndexDescriptor,
-	evalCtx *tree.EvalContext,
-) uint64 {
-	if c == nil || c.IsContradiction() || c.IsUnconstrained() {
-		return 0
-	}
-	if !i.Unique {
-		return 0
-	}
-
-	if c.Columns.Count() < len(i.ColumnIDs) {
-		// Not all cols specified.
-		return 0
-	}
-
-	var (
-		indexCols   opt.ColSet
-		notNullCols opt.ColSet
-	)
-	for _, id := range i.ColumnIDs {
-		if col, err := t.FindColumnByID(id); err != nil {
-			return 0
-		} else if !col.IsNullable() {
-			notNullCols.Add(opt.ColumnID(id))
-		}
-
-		indexCols.Add(opt.ColumnID(id))
-	}
-
-	return c.CalculateMaxResults(evalCtx, indexCols, notNullCols)
-}
-
-type indexInfo struct {
-	desc        *sqlbase.ImmutableTableDescriptor
-	index       *sqlbase.IndexDescriptor
-	cost        float64
-	covering    bool // Does the index cover the required IndexedVars?
-	reverse     bool
-	exactPrefix int
-
-	ic idxconstraint.Instance
-}
-
-func (v *indexInfo) init(s *scanNode) {
-	v.covering = v.isCoveringIndex(s)
-
-	// The base cost is the number of keys per row.
-	if v.index == &v.desc.PrimaryIndex {
-		// The primary index contains 1 key per column plus the sentinel key per
-		// row.
-		v.cost = float64(1 + len(v.desc.Columns) - len(v.desc.PrimaryIndex.ColumnIDs))
-	} else {
-		v.cost = 1
-		if !v.covering {
-			v.cost += float64(1 + len(v.desc.Columns) - len(v.desc.PrimaryIndex.ColumnIDs))
-			// Non-covering indexes are significantly more expensive than covering
-			// indexes.
-			v.cost *= nonCoveringIndexPenalty
-		}
-	}
-}
-
-// analyzeOrdering analyzes the ordering provided by the index and determines
-// if it matches the ordering requested by the query. Non-matching orderings
-// increase the cost of using the index.
-//
-// If preferOrderMatching is true, we prefer an index that matches the desired
-// ordering completely, even if it is not a covering index.
-func (v *indexInfo) analyzeOrdering(
-	ctx context.Context,
-	scan *scanNode,
-	analyzeOrdering analyzeOrderingFn,
-	preferOrderMatching bool,
-	evalCtx *tree.EvalContext,
-) {
-	// Analyze the ordering provided by the index (either forward or reverse).
-	fwdIndexProps := scan.computePhysicalProps(v.index, v.exactPrefix, false, evalCtx)
-	revIndexProps := scan.computePhysicalProps(v.index, v.exactPrefix, true, evalCtx)
-	fwdMatch, fwdOrderCols := analyzeOrdering(fwdIndexProps)
-	revMatch, revOrderCols := analyzeOrdering(revIndexProps)
-
-	if fwdOrderCols != revOrderCols {
-		panic(fmt.Sprintf("fwdOrderCols(%d) != revOrderCols(%d)", fwdOrderCols, revOrderCols))
-	}
-
-	orderCols := fwdOrderCols
-
-	// Weigh the cost by how much of the ordering matched.
-	//
-	// TODO(pmattis): Need to determine the relative weight for index selection
-	// based on sorting vs index selection based on filtering. Sorting is
-	// expensive due to the need to buffer up the rows and perform the sort, but
-	// not filtering is also expensive due to the larger number of rows scanned.
-	match := fwdMatch
-	if match < revMatch {
-		match = revMatch
-		v.reverse = true
-	}
-	weight := float64(orderCols+1) / float64(match+1)
-	v.cost *= weight
-
-	if match == orderCols && preferOrderMatching {
-		// Offset the non-covering index cost penalty.
-		v.cost *= (1.0 / nonCoveringIndexPenalty)
-	}
-
-	if log.V(2) {
-		log.Infof(ctx, "%s: analyzeOrdering: weight=%0.2f reverse=%v match=%d",
-			v.index.Name, weight, v.reverse, match)
-	}
-}
-
-// isCoveringIndex returns true if all of the columns needed from the scanNode are contained within
-// the index. This allows a scan of only the index to be performed without requiring subsequent
-// lookup of the full row.
-func (v *indexInfo) isCoveringIndex(scan *scanNode) bool {
-	if v.index.Type == sqlbase.IndexDescriptor_INVERTED {
-		return false
-	}
-	if v.index == &v.desc.PrimaryIndex {
-		// The primary key index always covers all of the columns.
-		return true
-	}
-
-	for _, colIdx := range scan.valNeededForCol.Ordered() {
-		// This is possible during a schema change when we have
-		// additional mutation columns.
-		if colIdx >= len(v.desc.Columns) && len(v.desc.MutationColumns()) > 0 {
-			return false
-		}
-		colID := v.desc.Columns[colIdx].ID
-		if !v.index.ContainsColumnID(colID) {
-			return false
-		}
-	}
-	return true
-}
-
-type indexInfoByCost []*indexInfo
-
-func (v indexInfoByCost) Len() int {
-	return len(v)
-}
-
-func (v indexInfoByCost) Less(i, j int) bool {
-	return v[i].cost < v[j].cost
-}
-
-func (v indexInfoByCost) Swap(i, j int) {
-	v[i], v[j] = v[j], v[i]
-}
-
-func (v indexInfoByCost) Sort() {
-	sort.Sort(v)
-}
-
-// makeIndexConstraints uses the opt code to generate index
-// constraints. Initializes v.ic, as well as v.exactPrefix and v.cost (with a
-// baseline cost for the index).
-func (v *indexInfo) makeIndexConstraints(
-	optimizer *xform.Optimizer, filters memo.FiltersExpr, evalCtx *tree.EvalContext,
-) error {
-	numIndexCols := len(v.index.ColumnIDs)
-
-	numExtraCols := 0
-	isInverted := (v.index.Type == sqlbase.IndexDescriptor_INVERTED)
-	// TODO(radu): we currently don't support index constraints on PK
-	// columns on an inverted index.
-	if !isInverted && !v.index.Unique {
-		// We have a non-unique index; the extra columns are added to the key and we
-		// can use them for index constraints.
-		numExtraCols = len(v.index.ExtraColumnIDs)
-	}
-
-	colIdxMap := make(map[sqlbase.ColumnID]int, len(v.desc.Columns))
-	for i := range v.desc.Columns {
-		colIdxMap[v.desc.Columns[i].ID] = i
-	}
-
-	columns := make([]opt.OrderingColumn, 0, numIndexCols+numExtraCols)
-	var notNullCols opt.ColSet
-	for i := 0; i < numIndexCols+numExtraCols; i++ {
-		var colID sqlbase.ColumnID
-		var dir encoding.Direction
-
-		if i < numIndexCols {
-			colID = v.index.ColumnIDs[i]
-			var err error
-			dir, err = v.index.ColumnDirections[i].ToEncodingDirection()
-			if err != nil {
-				return err
-			}
-		} else {
-			colID = v.index.ExtraColumnIDs[i-numIndexCols]
-			// Extra columns are always ascending.
-			dir = encoding.Ascending
-		}
-
-		idx, ok := colIdxMap[colID]
-		if !ok {
-			// Inactive column.
-			break
-		}
-
-		col := opt.MakeOrderingColumn(opt.ColumnID(idx+1), dir == encoding.Descending)
-		columns = append(columns, col)
-		if !v.desc.Columns[idx].Nullable {
-			notNullCols.Add(opt.ColumnID(idx + 1))
-		}
-	}
-	v.ic.Init(filters, columns, notNullCols, isInverted, evalCtx, optimizer.Factory())
-	idxConstraint := v.ic.Constraint()
-	if idxConstraint.IsUnconstrained() {
-		// The index isn't being restricted at all, bump the cost significantly to
-		// make any index which does restrict the keys more desirable.
-		v.cost *= 1000
-	} else {
-		v.exactPrefix = idxConstraint.ExactPrefix(evalCtx)
-		// Find the number of columns that are restricted in all spans.
-		numCols := len(columns)
-		for i := 0; i < idxConstraint.Spans.Count(); i++ {
-			sp := idxConstraint.Spans.Get(i)
-			// Take the max between the length of the start values and the end
-			// values.
-			n := sp.StartKey().Length()
-			if e := sp.EndKey().Length(); n < e {
-				n = e
-			}
-			// Take the minimum n across all spans.
-			if numCols > n {
-				numCols = n
-			}
-		}
-		// Boost the cost by what fraction of columns have constraints. The higher
-		// the fraction, the smaller the cost.
-		v.cost *= float64((numIndexCols + numExtraCols)) / float64(numCols)
-	}
-	return nil
-}
 
 func unconstrainedSpans(
 	tableDesc *sqlbase.ImmutableTableDescriptor, index *sqlbase.IndexDescriptor, forDelete bool,
